Cleaner design, move tests to separate directory and namespace
